Some deduplication for testing code
