+++ PyRIGS/tests/pages.py
@@ -2,7 +2,7 @@ from pypom import Page, Region
 from selenium.webdriver.common.by import By
 from selenium.webdriver import Chrome
 from selenium.common.exceptions import NoSuchElementException
-
+from PyRIGS.tests import regions
 
 class BasePage(Page):
     form_items = {}
@@ -34,36 +34,18 @@ class FormPage(BasePage):
         self.driver.execute_script("Array.from(document.getElementsByTagName(\"input\")).forEach(function (el, ind, arr) { el.removeAttribute(\"required\")});")
         self.driver.execute_script("Array.from(document.getElementsByTagName(\"select\")).forEach(function (el, ind, arr) { el.removeAttribute(\"required\")});")
 
+    def submit(self):
+        previous_errors = self.errors
+        self.find_element(*self._submit_locator).click()
+        self.wait.until(lambda x: self.errors != previous_errors or self.success)
+
     @property
     def errors(self):
         try:
-            error_page = self.ErrorPage(self, self.find_element(*self._errors_selector))
+            error_page = regions.ErrorPage(self, self.find_element(*self._errors_selector))
             return error_page.errors
         except NoSuchElementException:
             return None
-
-    class ErrorPage(Region):
-        _error_item_selector = (By.CSS_SELECTOR, "dl>span")
-
-        class ErrorItem(Region):
-            _field_selector = (By.CSS_SELECTOR, "dt")
-            _error_selector = (By.CSS_SELECTOR, "dd>ul>li")
-
-            @property
-            def field_name(self):
-                return self.find_element(*self._field_selector).text
-
-            @property
-            def errors(self):
-                return [x.text for x in self.find_elements(*self._error_selector)]
-
-        @property
-        def errors(self):
-            error_items = [self.ErrorItem(self, x) for x in self.find_elements(*self._error_item_selector)]
-            errors = {}
-            for error in error_items:
-                errors[error.field_name] = error.errors
-            return errors
 
 
 class LoginPage(BasePage):
