+++ PyRIGS/tests/regions.py
@@ -131,3 +131,27 @@ class SingleSelectPicker(Region):
     def set_value(self, value):
         picker = Select(self.root)
         picker.select_by_visible_text(value)
+
+
+class ErrorPage(Region):
+            _error_item_selector = (By.CSS_SELECTOR, "dl>span")
+
+            class ErrorItem(Region):
+                _field_selector = (By.CSS_SELECTOR, "dt")
+                _error_selector = (By.CSS_SELECTOR, "dd>ul>li")
+
+                @property
+                def field_name(self):
+                    return self.find_element(*self._field_selector).text
+
+                @property
+                def errors(self):
+                    return [x.text for x in self.find_elements(*self._error_selector)]
+
+            @property
+            def errors(self):
+                error_items = [self.ErrorItem(self, x) for x in self.find_elements(*self._error_item_selector)]
+                errors = {}
+                for error in error_items:
+                    errors[error.field_name] = error.errors
+                return errors
